+++ changelog.md
@@ -1,2 +1,9 @@
 # V1.7.1
 Add state space module and include "zhnmat" library.
+
+# V1.7.2
+Change continuous state space module from SISO to a general model.
+Add 2 functions that user can set the initial value and get the output value
+ of StateSpace, TransferFcn and DiscreteTransferFcn.
+Add version.
+Fix some bugs in destructor of Simulator.